akaros: Reduce net's diff from Plan 9
The intent is to easily spot where we are different from Plan 9.  This
removes some trivial differences (comments) and makes CloseRead and
CloseWrite internal.

Also, the usage of fd.data.read vs fd.read in udpsock is interesting.
With Plan 9's version, net test hangs.  It's probably related to a
timeout, but I didn't dig into it.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>