akaros: Remove the open-coded syscalls from exec
We can use the syscall generators to avoid rolling our own for
SYS_proc_create and friends.

I kept the byte slices for the serialized argv/env to limit the
shenanigans.  That stuff is passed off to parlib already.

One thing to note is that any auto-generated syscall that takes a []byte
will get the length of the slice inserted into the syscall arguments.
This is already what we want, but it means that the //sys templates are
not the syscall ABI - they are the Go API to the low-level syscalls.

Also, some of the syscalls are named the same thing as the Go functional
API.  For example, syscall.Exec is the Go function.  syscall.exec is the
Go wrapper for the syscall itself.  (Note the capitalization).  We do
something similar with {C,c}hdir.

Although []bytes are auto-converted into the pointer and the length,
this does not happen for strings.  That's why we have a bunch of
wrappers that add the string's length to the syscall parameters.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>