akaros: Add string lengths to autogenerated syscalls
We autogenerate the Go functions that make syscalls with the //sys
directive.  Those functions look like Go prototypes, not the kernel ABI.
The autogenerated magic handles the syscall struct, errno/errstr
conversions, and other stuff.

The //sys directive's parameters were never 1:1 with the syscall ABI.
Notably, byte slices added an extra parameter, even for other OSes.  For
our port, that meant that some calls were adding extra parameters, as
pointed out in previous commits for Getcwd.  In this commit, Readlink
was also doing that.

Now we extend that same behavior for []byte to strings, which is an
Akaros-specific treatment of strings.  All strings that get passed to
the kernel ABI must come with the string length.

This cleans up a bunch of wrappers, such as Rename and Unlink, including
their error handling.  Some syscalls were supported incorrectly and
would fail, such as Access.  That was in the 'not ported' section, but
we do have SYS_access.  Certain other wrappers must stick around.  For
instance, our SYS_chdir takes the target's pid.  Other functions simply
need to be ported completely, such as Mount.  Our mount takes an FD, not
a path.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>